+++ osznconnection-file-handling/src/main/java/org/complitex/osznconnection/file/entity/example/SubsidyExample.java
@@ -18,8 +18,8 @@ public class SubsidyExample extends AbstractRequestExample {
     private String building;
     private String corp;
     private String apartment;
-    private int start;
-    private int size;
+    private long start;
+    private long size;
     private String orderByClause;
     private boolean asc;
 
@@ -99,19 +99,19 @@ public class SubsidyExample extends AbstractRequestExample {
         this.apartment = apartment;
     }
 
-    public int getStart() {
+    public long getStart() {
         return start;
     }
 
-    public void setStart(int start) {
+    public void setStart(long start) {
         this.start = start;
     }
 
-    public int getSize() {
+    public long getSize() {
         return size;
     }
 
-    public void setSize(int size) {
+    public void setSize(long size) {
         this.size = size;
     }
